damage-removal: use new standard system
