+++ Blasphemous.LostDreams/Effects/IToggleEffect.cs
@@ -1,0 +1,7 @@
+﻿
+namespace Blasphemous.LostDreams.Effects;
+
+internal interface IToggleEffect
+{
+    public bool IsActive { get; }
+}
